// react-module/src/openapi/models/warehouse-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface WarehouseBasicDto
 */
export interface WarehouseBasicDto {
    /**
     * 
     * @type {number}
     * @memberof WarehouseBasicDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof WarehouseBasicDto
     */
    name?: any;
}